Linkedin_Post/requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
python-dotenv>=1.0.0